requirements.txt
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
pillow
numpy<2.0.0

--extra-index-url https://download.pytorch.org/whl/cu121
